requirements.txt
streamlit>=1.32.0
langchain>=0.1.16
langchain-google-genai>=0.0.8
google-generativeai>=0.8.5
python-dotenv>=1.0.1
